Moved Alinco DX-77-specific functions out of alinco.c and into dx77.c. The protocol used by the DX-77 is different than the DX-SR* line
